requirements.txt
Flask==2.3.3
requests==2.31.0
PyGithub==1.59.1
openai==0.28.1
python-dotenv==1.0.0
Werkzeug==2.3.7